JAMES-2813 move delete message vault query from webadmin to deleted-message-vault
