+++ mailbox/plugin/deleted-messages-vault/src/main/java/org/apache/james/vault/dto/query/QueryElement.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.webadmin.vault.routes.query;
+package org.apache.james.vault.dto.query;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
