+++ mailbox/plugin/deleted-messages-vault/src/test/java/org/apache/james/vault/dto/query/QueryElementDeserializerTest.java
@@ -17,14 +17,14 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.webadmin.vault.routes.query;
+package org.apache.james.vault.dto.query;
 
 import static org.apache.james.vault.DeletedMessageFixture.SUBJECT;
 import static org.apache.mailet.base.MailAddressFixture.SENDER;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import org.apache.james.webadmin.vault.routes.query.QueryTranslator.FieldName;
-import org.apache.james.webadmin.vault.routes.query.QueryTranslator.Operator;
+import org.apache.james.vault.dto.query.QueryTranslator.FieldName;
+import org.apache.james.vault.dto.query.QueryTranslator.Operator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
